project/requirements.txt
fastapi==0.88.0
uvicorn==0.22.0
langchain
langchain-nvidia-ai-endpoints
langchain-community
gradio==3.23.0
rich==13.3.0
pandas==1.5.3
faiss-cpu
numpy
python-dotenv
typing_extensions
langchain_core
urllib3
tqdm
charset_normalizer
idna
certifi
starlette
requests
tenacity
annotated_types
packaging
pydantic
sentence-transformers
